Only call GraphqlErrorBuilder with non-null data
The methods on GraphqlErrorBuilder each contain [an assertion that the supplied data is non-null](https://github.com/graphql-java/graphql-java/blob/master/src/main/java/graphql/GraphqlErrorBuilder.java#L77), but it is not required that these elements be non-null on every throw GraphQLError. For example, errors often omit a `path` property, which will lead to an unhandled `graphql.AssertException` when the error is handled by `GraphQLErrorFromExceptionHandler.java`.